# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Al Mehandi API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: orders
    description: Order management
  - name: admin
    description: Admin operations
  - name: qr
    description: QR code management
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /orders:
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  /orders/{orderId}/payment-screenshot:
    post:
      operationId: uploadPaymentScreenshot
      tags: [orders]
      summary: Upload payment screenshot for an order
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                screenshot:
                  type: string
                  format: binary
              required:
                - screenshot
      responses:
        "200":
          description: Screenshot uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLoginResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/logout:
    post:
      operationId: adminLogout
      tags: [admin]
      summary: Admin logout
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /admin/orders:
    get:
      operationId: adminGetOrders
      tags: [admin]
      summary: Get all orders (admin)
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"

  /admin/orders/{orderId}/status:
    patch:
      operationId: adminUpdateOrderStatus
      tags: [admin]
      summary: Update order status
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusRequest"
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  /admin/orders/export:
    get:
      operationId: adminExportOrders
      tags: [admin]
      summary: Export orders as CSV
      responses:
        "200":
          description: CSV export
          content:
            text/csv:
              schema:
                type: string

  /admin/qr/{comboId}:
    post:
      operationId: adminUploadQr
      tags: [qr]
      summary: Upload QR code for a combo
      parameters:
        - name: comboId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                qr:
                  type: string
                  format: binary
              required:
                - qr
      responses:
        "200":
          description: QR uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QrCode"

  /qr/{comboId}:
    get:
      operationId: getQrCode
      tags: [qr]
      summary: Get QR code for a combo
      parameters:
        - name: comboId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: QR code info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QrCode"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    CreateOrderRequest:
      type: object
      properties:
        comboId:
          type: integer
        comboName:
          type: string
        price:
          type: number
        fullName:
          type: string
        phone:
          type: string
        whatsapp:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        pincode:
          type: string
        notes:
          type: string
      required:
        - comboId
        - comboName
        - price
        - fullName
        - phone
        - whatsapp
        - address
        - city
        - state
        - pincode

    Order:
      type: object
      properties:
        id:
          type: integer
        comboId:
          type: integer
        comboName:
          type: string
        price:
          type: number
        fullName:
          type: string
        phone:
          type: string
        whatsapp:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        pincode:
          type: string
        notes:
          type: string
        status:
          type: string
        screenshotUrl:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - comboId
        - comboName
        - price
        - fullName
        - phone
        - whatsapp
        - address
        - city
        - state
        - pincode
        - status
        - createdAt

    UpdateOrderStatusRequest:
      type: object
      properties:
        status:
          type: string
          enum:
            - pending_verification
            - payment_verified
            - order_confirmed
            - cancelled
      required:
        - status

    AdminLoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password

    AdminLoginResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    QrCode:
      type: object
      properties:
        comboId:
          type: integer
        imageUrl:
          type: string
          nullable: true
      required:
        - comboId