# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Al Mehandi API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Create a new order
 */
export const CreateOrderBody = zod.object({
  comboId: zod.number(),
  comboName: zod.string(),
  price: zod.number(),
  fullName: zod.string(),
  phone: zod.string(),
  whatsapp: zod.string(),
  address: zod.string(),
  city: zod.string(),
  state: zod.string(),
  pincode: zod.string(),
  notes: zod.string().optional(),
});

/**
 * @summary Upload payment screenshot for an order
 */
export const UploadPaymentScreenshotParams = zod.object({
  orderId: zod.coerce.number(),
});

export const UploadPaymentScreenshotBody = zod.object({
  screenshot: zod.instanceof(File),
});

export const UploadPaymentScreenshotResponse = zod.object({
  id: zod.number(),
  comboId: zod.number(),
  comboName: zod.string(),
  price: zod.number(),
  fullName: zod.string(),
  phone: zod.string(),
  whatsapp: zod.string(),
  address: zod.string(),
  city: zod.string(),
  state: zod.string(),
  pincode: zod.string(),
  notes: zod.string().optional(),
  status: zod.string(),
  screenshotUrl: zod.string().nullish(),
  createdAt: zod.string(),
});

/**
 * @summary Admin login
 */
export const AdminLoginBody = zod.object({
  username: zod.string(),
  password: zod.string(),
});

export const AdminLoginResponse = zod.object({
  success: zod.boolean(),
  message: zod.string(),
});

/**
 * @summary Admin logout
 */
export const AdminLogoutResponse = zod.object({
  success: zod.boolean(),
  message: zod.string().optional(),
});

/**
 * @summary Get all orders (admin)
 */
export const AdminGetOrdersQueryParams = zod.object({
  search: zod.coerce.string().optional(),
  status: zod.coerce.string().optional(),
});

export const AdminGetOrdersResponseItem = zod.object({
  id: zod.number(),
  comboId: zod.number(),
  comboName: zod.string(),
  price: zod.number(),
  fullName: zod.string(),
  phone: zod.string(),
  whatsapp: zod.string(),
  address: zod.string(),
  city: zod.string(),
  state: zod.string(),
  pincode: zod.string(),
  notes: zod.string().optional(),
  status: zod.string(),
  screenshotUrl: zod.string().nullish(),
  createdAt: zod.string(),
});
export const AdminGetOrdersResponse = zod.array(AdminGetOrdersResponseItem);

/**
 * @summary Update order status
 */
export const AdminUpdateOrderStatusParams = zod.object({
  orderId: zod.coerce.number(),
});

export const AdminUpdateOrderStatusBody = zod.object({
  status: zod.enum([
    "pending_verification",
    "payment_verified",
    "order_confirmed",
    "cancelled",
  ]),
});

export const AdminUpdateOrderStatusResponse = zod.object({
  id: zod.number(),
  comboId: zod.number(),
  comboName: zod.string(),
  price: zod.number(),
  fullName: zod.string(),
  phone: zod.string(),
  whatsapp: zod.string(),
  address: zod.string(),
  city: zod.string(),
  state: zod.string(),
  pincode: zod.string(),
  notes: zod.string().optional(),
  status: zod.string(),
  screenshotUrl: zod.string().nullish(),
  createdAt: zod.string(),
});

/**
 * @summary Upload QR code for a combo
 */
export const AdminUploadQrParams = zod.object({
  comboId: zod.coerce.number(),
});

export const AdminUploadQrBody = zod.object({
  qr: zod.instanceof(File),
});

export const AdminUploadQrResponse = zod.object({
  comboId: zod.number(),
  imageUrl: zod.string().nullish(),
});

/**
 * @summary Get QR code for a combo
 */
export const GetQrCodeParams = zod.object({
  comboId: zod.coerce.number(),
});

export const GetQrCodeResponse = zod.object({
  comboId: zod.number(),
  imageUrl: zod.string().nullish(),
});