# lib/api-zod/src/generated/types/createOrderRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Al Mehandi API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateOrderRequest {
  comboId: number;
  comboName: string;
  price: number;
  fullName: string;
  phone: string;
  whatsapp: string;
  address: string;
  city: string;
  state: string;
  pincode: string;
  notes?: string;
}