# lib/api-zod/src/generated/types/qrCode.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Al Mehandi API specification
 * OpenAPI spec version: 0.1.0
 */

export interface QrCode {
  comboId: number;
  imageUrl?: string | null;
}